+++ SwiftCraftLauncher/MainFeature/Views/Toolbar/ContentToolbarView.swift
@@ -14,19 +14,22 @@ public struct ContentToolbarView: ToolbarContent {
     private var container
     @Environment(PlayerListViewModel.self)
     private var playerListViewModel
+    @Environment(GameRepository.self)
+    private var gameRepository
+
     @State private var showingAddPlayerSheet = false
     @State private var playerName = ""
     @State private var isPlayerNameValid = false
     @State private var showPlayerAlert = false
     @State private var showingGameForm = false
-    @Environment(GameRepository.self)
-    private var gameRepository
     @State private var showEditSkin = false
     @State private var showingMinecraftFriendsSheet = false
     @State private var minecraftFriendsSheetHost: MinecraftFriendsSheetHostAdapter?
     @State private var isLoadingFriends = false
     @State private var viewModel = ContentToolbarViewModel()
-    @State private var minecraftFriendsSheetViewModel = MinecraftFriendsSheetViewModel(friendsService: DIContainer.shared.ui.minecraftFriendsService)
+    @State private var minecraftFriendsSheetViewModel = MinecraftFriendsSheetViewModel(
+        friendsService: DIContainer.shared.ui.minecraftFriendsService,
+    )
 
     private var currentPlayer: Player? {
         playerListViewModel.currentPlayer
@@ -38,156 +41,35 @@ public struct ContentToolbarView: ToolbarContent {
 
     public var body: some ToolbarContent {
         ToolbarItemGroup(placement: .primaryAction) {
-            Button {
-                if currentPlayer == nil {
-                    showPlayerAlert = true
-                } else {
-                    showingGameForm.toggle()
-                }
-            } label: {
-                Label("game.form.title".localized(), systemImage: "plus")
-            }
-            .help("game.form.title".localized())
-            .sheet(isPresented: $showingGameForm) {
-                GameFormView()
-                    .presentationBackgroundInteraction(.automatic)
-            }
-            Spacer()
-            Button {
-                playerName = ""
-                isPlayerNameValid = false
-                showingAddPlayerSheet = true
-            } label: {
-                Label("player.add".localized(), systemImage: "person.badge.plus")
-            }
-            .help("player.add".localized())
-            .sheet(isPresented: $showingAddPlayerSheet) {
-                AddPlayerSheetView(
-                    playerName: $playerName,
-                    isPlayerNameValid: $isPlayerNameValid,
-                    onAdd: {
-                        if playerListViewModel.addPlayer(name: playerName) {
-                            AppLog.main.debug("Player \(playerName) added successfully (via ViewModel).")
-                        } else {
-                            AppLog.main.debug("Failed to add player \(playerName) (via ViewModel).")
-                        }
-                        isPlayerNameValid = true
-                        showingAddPlayerSheet = false
-                    },
-                    onCancel: {
-                        playerName = ""
-                        isPlayerNameValid = false
+            AddGameToolbarButton(
+                showingGameForm: $showingGameForm,
+                showPlayerAlert: $showPlayerAlert,
+                currentPlayer: currentPlayer,
+            )
 
-                        showingAddPlayerSheet = false
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-                            container.system.minecraftAuthService.clearAuthenticationData()
-                        }
-                    },
-                    onLogin: { profile in
-                        AppLog.main.debug("Premium login successful, user: \(profile.name)")
-                        _ = playerListViewModel.addOnlinePlayer(profile: profile)
-                        container.system.premiumAccountFlagManager.setPremiumAccountAdded()
-                        showingAddPlayerSheet = false
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-                            container.system.minecraftAuthService.clearAuthenticationData()
-                        }
-                    },
-                    onYggdrasilLogin: { profile in
-                        AppLog.main.debug("Yggdrasil login successful, user: \(profile.name)")
-                        OfflineUserServerMap.setServer(profile)
-                        _ = playerListViewModel.addOnlinePlayer(profile: profile)
-                        showingAddPlayerSheet = false
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-                            container.system.yggdrasilAuthService.logout()
-                        }
-                    },
-                    playerListViewModel: playerListViewModel,
-                )
-            }
-            .alert(isPresented: $showPlayerAlert) {
-                Alert(
-                    title: Text("sidebar.alert.no_player.title".localized()),
-                    message: Text("sidebar.alert.no_player.message".localized()),
-                    dismissButton: .default(Text("common.confirm".localized())),
-                )
-            }
+            Spacer()
+
+            AddPlayerToolbarButton(
+                showingAddPlayerSheet: $showingAddPlayerSheet,
+                playerName: $playerName,
+                isPlayerNameValid: $isPlayerNameValid,
+            )
 
             if isCurrentPlayerOnline {
-                Button {
-                    Task {
-                        await viewModel.preloadSkinDataForManager(player: currentPlayer)
-                        if currentPlayer != nil, !Task.isCancelled {
-                            showEditSkin = true
-                        }
-                    }
-                } label: {
-                    if viewModel.isLoadingSkin {
-                        ProgressView()
-                            .controlSize(.small)
-                    } else {
-                        Label("skin.title".localized(), systemImage: "tshirt")
-                    }
-                }
-                .help("skin.title".localized())
-                .disabled(viewModel.isLoadingSkin)
-                .sheet(isPresented: $showEditSkin) {
-                    SkinToolDetailView(
-                        preloadedSkinInfo: viewModel.preloadedSkinInfo,
-                        preloadedProfile: viewModel.preloadedProfile,
-                    )
-                    .onDisappear {
-                        viewModel.clearPreloadedSkinData()
-                    }
-                }
+                SkinToolbarButton(
+                    showEditSkin: $showEditSkin,
+                    isLoadingSkin: viewModel.isLoadingSkin,
+                    currentPlayer: currentPlayer,
+                    viewModel: viewModel,
+                )
 
-                Button {
-                    Task {
-                        guard let p = currentPlayer else { return }
-                        isLoadingFriends = true
-                        defer { isLoadingFriends = false }
-                        let host = MinecraftFriendsSheetHostAdapter(player: p)
-                        minecraftFriendsSheetHost = host
-                        minecraftFriendsSheetViewModel.prepare(playerId: p.id, host: host)
-                        await minecraftFriendsSheetViewModel.load(forceRefresh: false)
-                        guard !Task.isCancelled else { return }
-                        showingMinecraftFriendsSheet = true
-                    }
-                } label: {
-                    if isLoadingFriends {
-                        ProgressView()
-                            .controlSize(.small)
-                    } else {
-                        Label("minecraft.friends.toolbar.title".localized(), systemImage: "person.2")
-                    }
-                }
-                .help("minecraft.friends.toolbar.help".localized())
-                .disabled(isLoadingFriends)
-                .sheet(isPresented: $showingMinecraftFriendsSheet) {
-                    if let p = currentPlayer, minecraftFriendsSheetHost != nil {
-                        MinecraftFriendsSheetView(
-                            playerId: p.id,
-                            viewModel: minecraftFriendsSheetViewModel,
-                            localize: MinecraftFriendsSheetLocalize.resolver(
-                                localeIdentifier: { container.ui.languageManager.selectedLanguage },
-                                fallback: { $0.localized() },
-                            ),
-                            limitBodyScrollHeight: container.ui.generalSettingsManager.limitCommonSheetHeight,
-                        ) { _, url in
-                            Group {
-                                if let u = url, !u.isEmpty {
-                                    MinecraftSkinUtils(type: .url, src: u, size: 40)
-                                } else {
-                                    ProgressView()
-                                        .controlSize(.small)
-                                        .frame(width: 40, height: 40)
-                                }
-                            }
-                        }
-                        .onDisappear {
-                            minecraftFriendsSheetHost = nil
-                        }
-                    }
-                }
+                MinecraftFriendsToolbarButton(
+                    showingMinecraftFriendsSheet: $showingMinecraftFriendsSheet,
+                    minecraftFriendsSheetHost: $minecraftFriendsSheetHost,
+                    isLoadingFriends: $isLoadingFriends,
+                    currentPlayer: currentPlayer,
+                    viewModel: minecraftFriendsSheetViewModel,
+                )
             }
         }
     }
